requirements.txt
#requirements.txt

python-dotenv
pandas
gspread
oauth2client